+++ src/static_data/map_index.py
@@ -1,0 +1,212 @@
+"""
+Information about maps.
+
+Manually compiled.
+"""
+
+MAP_INDEX = {
+    # Display names are as per Steam app/workshop titles/text except where noted.
+    "abramia": {  # mod
+        "display_name": "Isla Abramia",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=648775794",
+    },
+    "Altis": {  # base game
+        "display_name": "Altis",
+        "url": "https://store.steampowered.com/app/107410",
+    },
+    "brf_sumava": {"display_name": "Šumava"},
+    "blud_vidda": {  # mod  # TODO: verify in-game
+        "display_name": r"Vidda \| legacy version",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1282716647",
+    },
+    "cam_lao_nam": {  # creator DLC: S.O.G. Prairie Fire
+        "display_name": "Cam Lao Nam",
+        "url": "https://store.steampowered.com/app/1227700",
+    },
+    "chernarus": {  # mod: CUP Terrains - Maps
+        "display_name": "Chernarus (Autumn)",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=583544987",
+    },
+    "chernarus_summer": {  # mod: CUP Terrains - Maps
+        "display_name": "Chernarus (Summer)",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=583544987",
+    },
+    "chernarus_winter": {  # mod: CUP Terrains - Maps
+        "display_name": "Chernarus (Winter)",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=583544987",
+    },
+    "chernarusredux": {"display_name": "Chernarus Redux"},
+    "cup_chernarus_A3": {  # mod: CUP Terrains - Maps 2.0
+        "display_name": "Chernarus 2020",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1981964169",
+    },
+    "Enoch": {  # DLC: Contact
+        "display_name": "Livonia",
+        "url": "https://store.steampowered.com/app/1021790",
+    },
+    "Esseker": {  # mod TODO: Verify in-game
+        "display_name": "New Esseker",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2712204589",
+    },
+    "fapovo": {  # mod
+        "display_name": "Fapovo",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1910457930",
+    },
+    "gm_weferlingen_summer": {  # CDLC: Global Mobilization - Cold War Germany
+        "display_name": "Weferlingen (Summer)",
+        "url": "https://store.steampowered.com/app/1042220",
+    },
+    "gm_weferlingen_winter": {  # CDLC: Global Mobilization - Cold War Germany
+        "display_name": "Weferlingen (Winter)",
+        "url": "https://store.steampowered.com/app/1042220",
+    },
+    "green_sea": {  # mod
+        "display_name": "Green Sea",  # officially "Green Sea Terrain"
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2645015212",
+    },
+    "Gulfcoast": {"display_name": "Gulfcoast Islands"},  # sic
+    "iron_excelsior_Tobruk": {  # mod: IFA3 AIO
+        "display_name": "Tobruk",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2648308937",
+    },
+    "isladuala3": {  # mod
+        "display_name": "Isladuala3",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=714149065",
+    },
+    "Kapaulio": {  # mod
+        "display_name": "Saint Kapaulio",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=939686262",
+    },
+    "Kunduz": {  # mod
+        "display_name": "Kunduz, Afghanistan",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=421620913",
+    },
+    "lingor3": {  # mod
+        "display_name": "Lingor Island",  # TODO: Dingor also?
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=718649903",
+    },
+    "Lythium": {  # mod
+        "display_name": "Lythium",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=909547724",
+    },
+    "Malden": {  # free DLC
+        "display_name": "Malden 2035",
+        "url": "https://store.steampowered.com/app/639600",
+    },
+    "mehland": {  # mod
+        "display_name": "Mehland",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=3317364155",
+    },
+    "namalsk": {  # mod
+        "display_name": "Namalsk [R]",  # aka Namalsk Revamped
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=3013966707",
+    },
+    "Napf": {  # mod
+        "display_name": "Napf Island A3",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1425442364",
+    },
+    "NapfWinter": {  # mod
+        "display_name": "Napf Island A3 (Winter)",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1425442364",
+    },
+    "OPTRE_Madrigal": {  # mod: Operation TREBUCHET
+        "display_name": "Madrigal",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=769440155",
+    },
+    "Panthera3": {  # mod  # TODO: inc Winter?
+        "display_name": "Island Panthera",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=708278910",
+    },
+    "pja310": {  # mod
+        "display_name": "G.O.S Al Rayak",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=648172507",
+    },
+    "psyfx_pht": {  # mod: Unsung
+        "display_name": "Phuoc Tuy",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=943001311",
+    },
+    "pulau": {  # mod
+        "display_name": "Pulau",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1423583812",
+    },
+    "rhspkl": {  # mod
+        "display_name": "Prei Khmaoch Luong",  # aka RHSPKL
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1978754337",
+    },
+    "ruha": {  # mod
+        "display_name": "Ruha",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1368857262",
+    },
+    "sara": {"display_name": "Sahrani"},
+    "SefrouRamal": {  # Creator DLC: Western Sahara
+        "display_name": "Sefrou-Ramal",
+        "url": "https://store.steampowered.com/app/1681170",
+    },
+    "Sehreno": {  # mod
+        "display_name": "Sehreno",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2876540929",
+    },
+    "SPE_Mortain": {  # Creator DLC: Spearhead 1994
+        "display_name": "Mortain",
+        "url": "https://store.steampowered.com/app/1175380",
+    },
+    "SPE_Normandy": {  # Creator DLC: Spearhead 1994
+        "display_name": "Normandy",
+        "url": "https://store.steampowered.com/app/1175380",
+    },
+    "spex_utah_beach": {  # requires Creator DLC: Spearhead 1944
+        "display_name": "Utah Beach",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=3117898674",
+    },
+    "Staszow": {"display_name": "Staszow"},
+    "StaszowWinter": {"display_name": "Staszow (Winter)"},
+    "stozec": {  # Creator DLC: CSLA Iron Curtain
+        "display_name": "Gabreta",
+        "url": "https://store.steampowered.com/app/1294440",
+    },
+    "Stratis": {  # base game
+        "display_name": "Stratis",
+        "url": "https://store.steampowered.com/app/107410",
+    },
+    "stubbhult": {
+        "display_name": "Stubbhult",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=3353189981",
+    },
+    "takistan": {  # mod: CUP Terrains - Maps
+        "display_name": "Takistan",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=583544987",
+    },
+    "Tanoa": {  # DLC: Apex
+        "display_name": "Tanoa",
+        "url": "https://store.steampowered.com/app/395180",
+    },
+    "tem_anizay": {  # mod
+        "display_name": "Anizay",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1537973181",
+    },
+    "tem_chernarus": {"display_name": "Svartmarka (Summer)"},
+    "tem_chernarusw": {"display_name": "Svartmarka (Winter)"},
+    "tem_kujari": {"display_name": "Kujari"},
+    "Tembelan": {"display_name": "Tembelan Island"},
+    "UMB_Colombia": {  # mod
+        "display_name": "UMB Colombia",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2266710560",
+    },
+    "vn_khe_sanh": {  # CDLC: SOG Prairie Fire
+        "display_name": "Khe Sanh",
+        "url": "https://store.steampowered.com/app/1227700",
+    },
+    "vt7": {  # mod
+        "display_name": "Virolahti - Valtatie 7",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=1926513010",
+    },
+    "Winthera3": {"display_name": "Island Winthera"},
+    "WW2_Omaha_Beach": {  # mod
+        "display_name": "Omaha Beach",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2993515557",
+    },
+    "yulakia": {  # mod
+        "display_name": "Yulakia",
+        "url": "https://steamcommunity.com/sharedfiles/filedetails/?id=2950257727",
+    },
+}